Move enums out of helper package
