+++ Kitodo-DataManagement/src/main/java/org/kitodo/data/database/beans/Folder.java
@@ -30,7 +30,7 @@ import javax.persistence.Table;
 import javax.persistence.Transient;
 
 import org.kitodo.api.imagemanagement.ImageManagementInterface;
-import org.kitodo.data.database.helper.enums.LinkingMode;
+import org.kitodo.data.database.enums.LinkingMode;
 import org.kitodo.forms.FolderGenerator;
 
 /**
