+++ Kitodo-DataManagement/src/main/java/org/kitodo/data/database/enums/LinkingMode.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package org.kitodo.data.database.helper.enums;
+package org.kitodo.data.database.enums;
 
 /**
  * Different ways how to link the contents of a folder in a METS fileGrp.
