+++ Kitodo-DataManagement/src/main/java/org/kitodo/data/database/enums/WorkflowStatus.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package org.kitodo.data.database.helper.enums;
+package org.kitodo.data.database.enums;
 
 /**
  * Enum for workflow status. Statuses:
