+++ Kitodo/src/main/java/org/kitodo/production/services/data/ProcessService.java
@@ -83,12 +83,12 @@ import org.kitodo.data.database.beans.Project;
 import org.kitodo.data.database.beans.Property;
 import org.kitodo.data.database.beans.Task;
 import org.kitodo.data.database.beans.User;
+import org.kitodo.data.database.enums.BatchType;
+import org.kitodo.data.database.enums.IndexAction;
+import org.kitodo.data.database.enums.LinkingMode;
+import org.kitodo.data.database.enums.MetadataFormat;
+import org.kitodo.data.database.enums.TaskStatus;
 import org.kitodo.data.database.exceptions.DAOException;
-import org.kitodo.data.database.helper.enums.BatchType;
-import org.kitodo.data.database.helper.enums.IndexAction;
-import org.kitodo.data.database.helper.enums.LinkingMode;
-import org.kitodo.data.database.helper.enums.MetadataFormat;
-import org.kitodo.data.database.helper.enums.TaskStatus;
 import org.kitodo.data.database.persistence.ProcessDAO;
 import org.kitodo.data.elasticsearch.exceptions.CustomResponseException;
 import org.kitodo.data.elasticsearch.index.Indexer;
