+++ Kitodo/src/main/java/org/kitodo/production/services/data/WorkflowService.java
@@ -19,8 +19,8 @@ import java.util.Objects;
 import org.elasticsearch.index.query.BoolQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.kitodo.data.database.beans.Workflow;
+import org.kitodo.data.database.enums.WorkflowStatus;
 import org.kitodo.data.database.exceptions.DAOException;
-import org.kitodo.data.database.helper.enums.WorkflowStatus;
 import org.kitodo.data.database.persistence.WorkflowDAO;
 import org.kitodo.data.elasticsearch.index.Indexer;
 import org.kitodo.data.elasticsearch.index.type.WorkflowType;
