+++ Kitodo/src/main/java/org/kitodo/production/services/schema/SchemaService.java
@@ -23,8 +23,8 @@ import org.kitodo.api.dataformat.MediaUnit;
 import org.kitodo.api.dataformat.MediaVariant;
 import org.kitodo.api.dataformat.Workpiece;
 import org.kitodo.data.database.beans.Folder;
-import org.kitodo.data.database.helper.enums.LinkingMode;
 import org.kitodo.data.database.beans.Process;
+import org.kitodo.data.database.enums.LinkingMode;
 import org.kitodo.export.ExportMets;
 import org.kitodo.production.helper.VariableReplacer;
 import org.kitodo.production.helper.metadata.legacytypeimplementations.LegacyInnerPhysicalDocStructHelper;
