+++ Kitodo/src/test/java/org/kitodo/MockDatabase.java
@@ -55,8 +55,6 @@ import org.kitodo.data.database.beans.Filter;
 import org.kitodo.data.database.beans.Folder;
 import org.kitodo.data.database.beans.LdapGroup;
 import org.kitodo.data.database.beans.LdapServer;
-import org.kitodo.data.database.helper.enums.BatchType;
-import org.kitodo.data.database.helper.enums.LinkingMode;
 import org.kitodo.data.database.beans.ListColumn;
 import org.kitodo.data.database.beans.Process;
 import org.kitodo.data.database.beans.Project;
@@ -67,12 +65,14 @@ import org.kitodo.data.database.beans.Task;
 import org.kitodo.data.database.beans.Template;
 import org.kitodo.data.database.beans.User;
 import org.kitodo.data.database.beans.Workflow;
+import org.kitodo.data.database.enums.BatchType;
+import org.kitodo.data.database.enums.LinkingMode;
+import org.kitodo.data.database.enums.PasswordEncryption;
+import org.kitodo.data.database.enums.PropertyType;
+import org.kitodo.data.database.enums.TaskEditType;
+import org.kitodo.data.database.enums.TaskStatus;
+import org.kitodo.data.database.enums.WorkflowStatus;
 import org.kitodo.data.database.exceptions.DAOException;
-import org.kitodo.data.database.helper.enums.PasswordEncryption;
-import org.kitodo.data.database.helper.enums.PropertyType;
-import org.kitodo.data.database.helper.enums.TaskEditType;
-import org.kitodo.data.database.helper.enums.TaskStatus;
-import org.kitodo.data.database.helper.enums.WorkflowStatus;
 import org.kitodo.data.database.persistence.HibernateUtil;
 import org.kitodo.data.elasticsearch.index.IndexRestClient;
 import org.kitodo.data.exceptions.DataException;
